% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc.R
\name{bc}
\alias{bc}
\title{Bayesian Combination}
\usage{
bc(grd, obs, nm)
}
\arguments{
\item{grd}{is a SpatialPointsDataFrame of gridded rainfall data.}

\item{obs}{is a SpatialPointsDataFrame of observed rainfall data}

\item{nm}{names of series to utilise}
}
\value{
A SpatialPointsDataFrame of merged fields of the same dimension as grd, unless \code{only_obs == TRUE} in which case the columns are determined by the intersection
}
\description{
This function merges satellite and Ordinary-Kriged rain
gauge fields using Bayesian inference
References: Todini, E. (2001). A Bayesian technique for
conditioning radar precipitation estimates to rain-gauge measurements.
Hydrology and Earth System Sciences, 5(2):187-199.
PROGEA Srl (2009). RAINMUSIC, User manual & references. PROGEA Srl, Bologna.
}
