% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine}
\alias{combine}
\title{Merge Series and observations}
\usage{
combine(
  grd,
  obs,
  method = c("dk", "ok", "ked", "mbc", "ridw", "bc"),
  nm = names(grd)
)
}
\arguments{
\item{grd}{is a sf object of gridded rainfall data.}

\item{obs}{is a sf object of observed rainfall data}

\item{method}{combination method (see details)}

\item{nm}{names of series to evaluate}
}
\value{
A sf of merged fields of the same dimension as grd, unless \code{only_obs == TRUE} in which case the columns are determined by the intersection of the names of grd and obs.
}
\description{
This function merges gridded fields and gauge data. A number of techniques are available: Double Kernal Smoothing (dk); Kriging with drift (ked); Bayesian combination (bc); multiplicative bias correction (mbc) and inverse distance weighting of the residuals (ridw). Ordinary Kriging (ok) is also provided for the interpolation of gauge data. See details for more informations about the methods.
}
\details{
the double kernel smoothing of the residuals following Li & Shao (2010)
References: Li, M. and Shao, Q. (2010).
An improved statistical approach to merge satellite rainfall
estimates and raingauge data. Journal of Hydrology, 385(1-4):51-64.
Silverman, B. (1986). Density Estimation for Statistics and Data Analysis.
Chapman and Hall/CRC Monographs on Statistics and Applied Probability Series.
Chapman and Hall/CRC.
}
\examples{
# out <- combine(sat, obs,method="dk")

}
