% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dk.R
\name{dk}
\alias{dk}
\title{Double Kernel Residual Smoothing}
\usage{
dk(rst, obs_location, obs, name_attr = "name")
}
\arguments{
\item{rst}{is a SpatRaster of gridded rainfall data.}

\item{obs_location}{is a SpatVector of observation locations.}

\item{obs}{is a zoo object of observations}

\item{name_attr}{the column of attributes in obs_location used to name the columns in obs}
}
\value{
A SpatRaster of merged fields of the same dimension as rst
}
\description{
This function merges gridded and rain gauge fields using
the double kernel smoothing of the residuals following Li & Shao (2010)
References: Li, M. and Shao, Q. (2010).
An improved statistical approach to merge satellite rainfall
estimates and raingauge data. Journal of Hydrology, 385(1-4):51-64.
Silverman, B. (1986). Density Estimation for Statistics and Data Analysis.
Chapman and Hall/CRC Monographs on Statistics and Applied Probability Series.
Chapman and Hall/CRC.
}
\examples{
# out <- dk(sat, gauge_loc, gauge)

}
