% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ked.R
\name{ked}
\alias{ked}
\title{Kriging with External Drift}
\usage{
ked(grd, obs, nm)
}
\arguments{
\item{grd}{is a SpatialPointsDataFrame of gridded rainfall data.}

\item{obs}{is a SpatialPointsDataFrame of observed rainfall data}

\item{nm}{names of series to utilise}
}
\value{
A SpatialPointsDataFrame of merged fields of the same dimension as grd, unless \code{only_obs == TRUE} in which case the columns are determined by the intersection of the names of grd and obs.
}
\description{
This function performs a Kriging interpolation of rain gauge
estimates using gridded estimates as an external drift
}
\examples{
# KED_out <- KED(sat, gauge)
# KED_cv  <- KED(sat, gauge, cross.val=TRUE)

}
