% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbc.R
\name{mbc}
\alias{mbc}
\title{Mean Bias Correction}
\usage{
mbc(grd, obs, nm)
}
\arguments{
\item{grd}{is a SpatialPointsDataFrame of gridded rainfall data.}

\item{obs}{is a SpatialPointsDataFrame of observed rainfall data}

\item{nm}{names of columns to evaluate}
}
\value{
A SpatialPointsDataFrame of merged fields of the same dimension as grd, unless \code{only_obs == TRUE} in which case the columns are determined by the intersection of the names of grd and obs#'
}
\description{
This function calculates a spatial mean multiplicative bias
correction of the satellite field at every time step
}
\examples{
# MBC_out <- MBC(sat, gauge)
# MBC_cv  <- MBC(sat, gauge, cross.val=TRUE)

}
