% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ok.R
\name{ok}
\alias{ok}
\title{Ordinary Kriging}
\usage{
ok(grd, obs)
}
\arguments{
\item{grd}{is a SpatialPoints object of predicted locations}

\item{obs}{is a SpatialPointsDataFrame of observed rainfall data}
}
\value{
A SpatialPointsDataFrame of Ordinary Kriged feilds. The number of points is setermined by grd with the number of series from obs. The Spatial Correlation is determined for each series seperately
}
\description{
Ordinary Kriging
}
\examples{
# out <- ok(sat, obs)

}
