% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIDW.R
\name{RIDW}
\alias{RIDW}
\title{Residual IDW interpolation}
\usage{
RIDW(sat, gauge, longlat = TRUE, cross.val = FALSE)
}
\arguments{
\item{sat}{is a list
sat[\link{2}] is a data.frame with dimensions nrow=no of satellite
pixels, containing columns x=x-coordinates, y=y-coordinates
sat  [\link{1}] is a zoo object with dimensions ncol=no of satellite
pixels, nrow=no of timestep}

\item{gauge}{is a list:
gauge[\link{2}] is a data.frame with dimensions nrow=no of stations,
containing columns x=x-coordinates, y=y-coordinates
gauge[\link{1}] is a zoo object with dimensions ncol=no of stations,
nrow=no of timestep}

\item{longlat}{is a flag to describe the coordinate grids of spatial data.
Defaults to TRUE i.e. on long-lat notation}

\item{cross.val}{option TRUE=in cross validation mode; default FALSE}
}
\value{
Zs merging output - zoo object with dimensions == dimensions sat
[\link{"ts""}

crossval cross validation- zoo object with dimensions == dimensions
gauge[\link{"ts""}
}
\description{
This function calculates the residual (additive bias) at every
pixel-point pair and interpolates them to the satellite field using inverse-
distance weighting (IDW) at every time step
}
\examples{
# RIDW_out <- RIDW(sat, gauge)
# RIDW_cv  <- RIDW(sat, gauge, cross.val=TRUE)

}
\author{
Bastian Manz
}
