% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dk.R
\name{dk}
\alias{dk}
\title{Double Kernel Residual Smoothing}
\usage{
dk(grd, obs, nm)
}
\arguments{
\item{grd}{is a sf of gridded rainfall data.}

\item{obs}{is a sf of observed rainfall data}

\item{nm}{names of series to utilise}
}
\value{
A sf of merged fields of the same dimension as grd, unless \code{only_obs == TRUE} in which case the columns are determined by the intersection of the names of grd and obs.
}
\description{
This function merges gridded and rain gauge fields using
the double kernel smoothing of the residuals following Li & Shao (2010)
References: Li, M. and Shao, Q. (2010).
An improved statistical approach to merge satellite rainfall
estimates and raingauge data. Journal of Hydrology, 385(1-4):51-64.
Silverman, B. (1986). Density Estimation for Statistics and Data Analysis.
Chapman and Hall/CRC Monographs on Statistics and Applied Probability Series.
Chapman and Hall/CRC.
}
\examples{
# out <- dk(sat, obs)

}
