% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridw.R
\name{ridw}
\alias{ridw}
\title{Residual IDW interpolation}
\usage{
ridw(grd, obs, nm)
}
\arguments{
\item{grd}{is a SpatialPointsDataFrame of gridded rainfall data.}

\item{obs}{is a SpatialPointsDataFrame of observed rainfall data}

\item{nm}{names of series to utilise}
}
\value{
A SpatialPointsDataFrame of merged fields of the same dimension as grd, unless \code{only_obs == TRUE} in which case the columns are determined by the intersection of the names of grd and obs.
}
\description{
This function calculates the residual (additive bias) at every
pixel-point pair and interpolates them to the satellite field using inverse-
distance weighting (IDW) at every time step
}
\examples{
# RIDW_out <- RIDW(sat, gauge)
# RIDW_cv  <- RIDW(sat, gauge, cross.val=TRUE)

}
\author{
Bastian Manz
}
